Refactor lightcurve.py into {lightcurve,lightcurvefile,correctors}.py
