+++ lightkurve/__init__.py
@@ -5,6 +5,8 @@ import os
 PACKAGEDIR = os.path.abspath(os.path.dirname(__file__))
 from .prf import *
 from .lightcurve import *
+from .lightcurvefile import *
+from .correctors import *
 from .targetpixelfile import *
 from .utils import *
 from .convenience import *
