+++ lightkurve/tests/test_lightcurve.py
@@ -4,17 +4,15 @@ import pytest
 import numpy as np
 from numpy.testing import (assert_almost_equal, assert_array_equal,
                            assert_allclose)
-from astropy.utils.data import get_pkg_data_filename
 from astropy.io import fits as pyfits
-from ..lightcurve import (LightCurve, KeplerCBVCorrector, KeplerLightCurveFile,
-                          TessLightCurveFile, SFFCorrector, KeplerLightCurve,
-                          TessLightCurve, iterative_box_period_search)
+from ..lightcurve import (LightCurve, KeplerLightCurve, iterative_box_period_search)
+from ..lightcurvefile import KeplerLightCurveFile, TessLightCurveFile
 
 # 8th Quarter of Tabby's star
 TABBY_Q8 = ("https://archive.stsci.edu/missions/kepler/lightcurves"
             "/0084/008462852/kplr008462852-2011073133259_llc.fits")
 TABBY_TPF = ("https://archive.stsci.edu/missions/kepler/target_pixel_files"
-            "/0084/008462852/kplr008462852-2011073133259_lpd-targ.fits.gz")
+             "/0084/008462852/kplr008462852-2011073133259_lpd-targ.fits.gz")
 K2_C08 = ("https://archive.stsci.edu/missions/k2/lightcurves/c8/"
           "220100000/39000/ktwo220139473-c08_llc.fits")
 KEPLER10 = ("https://archive.stsci.edu/missions/kepler/lightcurves/"
@@ -59,19 +57,8 @@ def test_rmath_operators():
 
 
 @pytest.mark.remote_data
-def test_kepler_cbv_fit():
-    # comparing that the two methods to do cbv fit are the nearly the same
-    cbv = KeplerCBVCorrector(TABBY_Q8)
-    cbv_lc = cbv.correct()
-    assert_almost_equal(cbv.coeffs, [0.08534423, 0.10814261], decimal=3)
-    lcf = KeplerLightCurveFile(TABBY_Q8)
-    cbv_lcf = lcf.compute_cotrended_lightcurve()
-    assert_almost_equal(cbv_lc.flux, cbv_lcf.flux)
-
-
-@pytest.mark.remote_data
 @pytest.mark.parametrize("path, mission", [(TABBY_Q8, "Kepler"), (K2_C08, "K2")])
-def test_KeplerLightCurve(path, mission):
+def test_KeplerLightCurveFile(path, mission):
     lcf = KeplerLightCurveFile(path, quality_bitmask=None)
     hdu = pyfits.open(path)
     kplc = lcf.get_lightcurve('SAP_FLUX')
@@ -113,9 +100,10 @@ def test_bitmasking(quality_bitmask, answer):
     flux = lcf.get_lightcurve('SAP_FLUX').flux
     assert len(flux) == answer
 
+
 def test_lightcurve_fold():
     """Test the ``LightCurve.fold()`` method."""
-    lc = LightCurve(time=np.linspace(0,10,100), flux=np.zeros(100)+1)
+    lc = LightCurve(time=np.linspace(0, 10, 100), flux=np.zeros(100)+1)
     fold = lc.fold(period=1)
     assert_almost_equal(fold.time[0], -0.5, 2)
     assert_almost_equal(np.min(fold.time), -0.5, 2)
@@ -149,6 +137,7 @@ def test_lightcurve_plot():
     lcf.plot()
     lcf.SAP_FLUX.plot()
 
+
 def test_cdpp():
     """Test the basics of the CDPP noise metric."""
     # A flat lightcurve should have a CDPP close to zero
@@ -166,54 +155,6 @@ def test_cdpp_tabby():
     lc = LightCurve(lcf.PDCSAP_FLUX.time[:1000], lcf.PDCSAP_FLUX.flux[:1000])
     assert(np.abs(lc.cdpp() - lcf.header(ext=1)['CDPP6_0']) < 30)
 
-def test_sff_corrector():
-    """Does our code agree with the example presented in Vanderburg
-    and Jhonson (2014)?"""
-    # The following csv file, provided by Vanderburg and Jhonson
-    # at https://www.cfa.harvard.edu/~avanderb/k2/ep60021426.html,
-    # contains the results of applying SFF to EPIC 60021426.
-    fn = get_pkg_data_filename('./data/ep60021426alldiagnostics.csv')
-    data = np.genfromtxt(fn, delimiter=',', skip_header=1)
-    mask = data[:, -2] == 0 # indicates whether the thrusters were on or off
-    time = data[:, 0]
-    raw_flux = data[:, 1]
-    corrected_flux = data[:, 2]
-    centroid_col = data[:, 3]
-    centroid_row = data[:, 4]
-    arclength = data[:, 5]
-    correction = data[:, 6]
-
-    sff = SFFCorrector()
-    corrected_lc = sff.correct(time=time, flux=raw_flux,
-                               centroid_col=centroid_col,
-                               centroid_row=centroid_row,
-                               niters=1)
-    # do hidden plots execute smoothly?
-    ax = sff._plot_rotated_centroids()
-    ax = sff._plot_normflux_arclength()
-
-    # the factor self.bspline(time-time[0]) accounts for
-    # the long term trend which is divided out in order to get a "flat"
-    # lightcurve.
-    assert_almost_equal(corrected_lc.flux*sff.bspline(time-time[0]),
-                        corrected_flux, decimal=3)
-    assert_array_equal(time, corrected_lc.time)
-    # the factor of 4 below accounts for the conversion
-    # between pixel units to arcseconds
-    assert_almost_equal(4*sff.s, arclength, decimal=2)
-    assert_almost_equal(sff.interp(sff.s), correction, decimal=3)
-
-    # test using KeplerLightCurve interface
-    klc = KeplerLightCurve(time=time, flux=raw_flux, centroid_col=centroid_col,
-                           centroid_row=centroid_row)
-    klc = klc.correct(niters=1)
-    sff = klc.corrector
-
-    assert_almost_equal(klc.flux*sff.bspline(time-time[0]),
-                        corrected_flux, decimal=3)
-    assert_almost_equal(4*sff.s, arclength, decimal=2)
-    assert_almost_equal(sff.interp(sff.s), correction, decimal=3)
-    assert_array_equal(time, klc.time)
 
 def test_bin():
     lc = LightCurve(time=np.arange(10), flux=2*np.ones(10),
@@ -222,6 +163,7 @@ def test_bin():
     assert_allclose(binned_lc.flux, 2*np.ones(5))
     assert_allclose(binned_lc.flux_err, np.ones(5))
     assert len(binned_lc.time) == 5
+
 
 def test_normalize():
     """Does the `LightCurve.normalize()` method normalize the flux?"""
@@ -232,7 +174,7 @@ def test_normalize():
 @pytest.mark.remote_data
 def test_iterative_box_period_search():
     """Can we recover the orbital period of Kepler-10b?"""
-    answer = 0.837495 # wikipedia
+    answer = 0.837495  # wikipedia
     klc = KeplerLightCurveFile(KEPLER10)
     pdc = klc.PDCSAP_FLUX
     flat, trend = pdc.flatten(return_trend=True)
@@ -255,6 +197,7 @@ def test_to_pandas():
         # pandas is an optional dependency
         pass
 
+
 def test_to_table():
     """Test the `LightCurve.to_table()` method."""
     time, flux, flux_err = range(3), np.ones(3), np.zeros(3)
@@ -263,6 +206,7 @@ def test_to_table():
     assert_allclose(tbl['time'], time)
     assert_allclose(tbl['flux'], flux)
     assert_allclose(tbl['flux_err'], flux_err)
+
 
 def test_to_csv():
     """Test the `LightCurve.to_csv()` method."""
@@ -273,6 +217,7 @@ def test_to_csv():
     except ImportError:
         # pandas is an optional dependency
         pass
+
 
 def test_date():
     '''Test the lc.date() function'''
